implemented textures handling in batch renderer.
#..............How textures are implemented in batch renderer?.................#
In renderer2D::init() function we pass the array of slots (Containing {0,1,2,3,.....32}) to a Sampler2D uniform array in the fragment Shader.

Then at each draw of quad with textures we bind that texture to a specified slot and that slot number becomes the index of the array (as both slots and indices starts from 0). That slot is passed as an vertex attribute in vertex shader.